extract classes in separate files
